requirements.txt
streamlit==1.27.0
rich==13.4.2
markdown-it-py==4.0.0
pygments==2.19.2
mdurl==0.1.2
Pillow==9.5.0
ultralytics==8.2.1
torch==2.2.0
torchvision==0.15.2
opencv-python-headless==4.8.1.78
numpy==1.26.0
